Fixed compilation warnings
git-svn-id: https://svn.pjsip.org/repos/pjproject/trunk@1418 74dad513-b988-da41-8d7b-12977e46ad98

+++ pjlib-util/src/pjlib-util-test/resolver_test.c
@@ -251,7 +251,7 @@ static int print_packet(const pj_dns_parsed_packet *rec, pj_uint8_t *pkt,
 #endif
 
     pj_assert(sizeof(pj_dns_hdr)==12);
-    if (size < sizeof(pj_dns_hdr))
+    if (size < (int)sizeof(pj_dns_hdr))
 	return -1;
 
     /* Initialize header */
